[issue-118] Add several config files
