+++ alertmgr/anonymizer.go
@@ -1,0 +1,26 @@
+package alertmgr
+
+import "reflect"
+
+func anonymizeSettings(settings *PluginSettings) *PluginSettings {
+	fieldsToAnonymize := [...]string{
+		"User",
+		"Password",
+		"Url",
+		"InstanceName",
+	}
+	copyToAnonymize := *settings
+
+	for _, key := range fieldsToAnonymize {
+
+		r := reflect.ValueOf(&copyToAnonymize)
+		v := reflect.Indirect(r).FieldByName(key)
+		prop := v.String()
+
+		if prop != "" {
+			v.SetString(AnonymizeReplacement)
+		}
+	}
+
+	return &copyToAnonymize
+}
