+++ main.go
@@ -6,26 +6,28 @@ import (
 	"github.com/aquasecurity/postee/utils"
 	"github.com/aquasecurity/postee/webserver"
 	"github.com/spf13/cobra"
+	"io/ioutil"
 	"log"
 	"os"
 	"os/signal"
 	"runtime"
+	"strings"
 	"syscall"
 )
 
 const (
-	URL       = "0.0.0.0:8082"
-	TLS       = "0.0.0.0:8445"
-	URL_USAGE = "The socket to bind to, specified using host:port."
-	TLS_USAGE = "The TLS socket to bind to, specified using host:port."
-	CFG_FILE  = "/config/cfg.yaml"
-	CFG_USAGE = "The alert configuration file."
+	URL        = "0.0.0.0:8082"
+	TLS        = "0.0.0.0:8445"
+	URL_USAGE  = "The socket to bind to, specified using host:port."
+	TLS_USAGE  = "The TLS socket to bind to, specified using host:port."
+	CFG_USAGE  = "The folder which contains alert configuration files."
+	CFG_FOLDER = "/config/"
 )
 
 var (
-	url     = ""
-	tls     = ""
-	cfgfile = ""
+	url       = ""
+	tls       = ""
+	cfgFolder = ""
 )
 
 var rootCmd = &cobra.Command{
@@ -37,7 +39,22 @@ var rootCmd = &cobra.Command{
 func init() {
 	rootCmd.Flags().StringVar(&url, "url", URL, URL_USAGE)
 	rootCmd.Flags().StringVar(&tls, "tls", TLS, TLS_USAGE)
-	rootCmd.Flags().StringVar(&cfgfile, "cfgfile", CFG_FILE, CFG_USAGE)
+	rootCmd.Flags().StringVar(&cfgFolder, "cfgFolder", CFG_FOLDER, CFG_USAGE)
+}
+
+func getFilesFromFolder(folder string) ([]string, error) {
+	files, err := ioutil.ReadDir(folder)
+	if err != nil {
+		return nil, err
+	}
+	names := []string{}
+	if !strings.HasSuffix(folder, "/") {
+		folder += "/"
+	}
+	for _, file := range files {
+		names = append(names, folder+file.Name())
+	}
+	return names, nil
 }
 
 func main() {
@@ -54,11 +71,18 @@ func main() {
 			tls = os.Getenv("AQUAALERT_TLS")
 		}
 
-		if os.Getenv("AQUAALERT_CFG") != "" {
-			cfgfile = os.Getenv("AQUAALERT_CFG")
+		cfgFolder = os.Getenv("AQUAALERT_CFG_FOLDER")
+		if cfgFolder == "" {
+			log.Printf("AQUAALERT_CFG_FOLDER environment variable is empty!")
+			return
+		}
+		files, err := getFilesFromFolder(cfgFolder)
+		if err != nil {
+			log.Printf("getFilesFromFolder(%q) error: %v", cfgFolder, err)
+			return
 		}
 
-		go alertmgr.Instance().Start(cfgfile)
+		go alertmgr.Instance().Start(files)
 		defer alertmgr.Instance().Terminate()
 
 		go webserver.Instance().Start(url, tls)
